feat: add repr json for logicalplan
