+++ daft/daft/__init__.pyi
@@ -1831,6 +1831,7 @@ class LogicalPlanBuilder:
     def to_adaptive_physical_plan_scheduler(self, cfg: PyDaftExecutionConfig) -> AdaptivePhysicalPlanScheduler: ...
     def repr_ascii(self, simple: bool) -> str: ...
     def repr_mermaid(self, options: MermaidOptions) -> str: ...
+    def repr_json(self, include_schema: bool) -> str: ...
 
 class DistributedPhysicalPlan:
     @staticmethod
